backend/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
numpy==1.22.3
psycopg2>=2.8
django-cors-headers
